chore: more refactoring, create parsers dir
